Update README.md
# Problem Staement 
In India, the industrial classification of the workforce is essential to understand the distribution of the labor force across various sectors. The classification of main workers and marginal workers, other than cultivators and agricultural laborers, by sex and by section, division, and class, has been traditionally used to understand the economic status and employment trends in the country. However, the current data on this classification is outdated and may not accurately reflect the current state of the workforce. The aim of this study is to update the information on the industrial classification of the main and marginal workers, other than cultivators and agricultural laborers, by sex and by section, division, and class, to provide relevant and accurate data for policy making and employment planning.